Rename banzai.celery -> banzai.Schedule
Resolves a naming conflict with the Celery library:

ImportError: cannot import name 'Celery' from 'celery' (consider renaming '/home/austin/Documents/lco/code/banzai/banzai/celery.py' if it has the same name as a library you intended to import)

If/when banzai moves away from Celery this will make more sense
naming-wise.
