+++ banzai/celeryconfig.py
@@ -1,5 +1,5 @@
 broker_url = 'redis://localhost:6379/0'
-imports = ('banzai.main', 'banzai.celery',)
+imports = ('banzai.main', 'banzai.scheduling',)
 worker_prefetch_multiplier = 1
 worker_max_tasks_per_child = 100
 task_acks_late = True
